+++ src/Listeners/CacheUsersWithMembershipToGroup.php
@@ -2,12 +2,13 @@
 
 namespace BristolSU\UnionCloud\Listeners;
 
-use BristolSU\ControlDB\Cache\DataUser;
 use BristolSU\ControlDB\Cache\Pivots\UserGroup;
+use BristolSU\ControlDB\Contracts\Models\DataUser;
+use BristolSU\ControlDB\Events\Pivots\UserGroup\UserAddedToGroup;
+use BristolSU\ControlDB\Events\Pivots\UserGroup\UserRemovedFromGroup;
+use BristolSU\ControlDB\Contracts\Repositories\User;
 use BristolSU\UnionCloud\Events\UsersWithMembershipToGroupRetrieved;
-use BristolSU\UnionCloud\Models\DataUserModel;
 use Illuminate\Contracts\Cache\Repository;
-use Illuminate\Support\Facades\Cache;
 
 class CacheUsersWithMembershipToGroup
 {
@@ -27,9 +28,26 @@ class CacheUsersWithMembershipToGroup
      */
     public function handle(UsersWithMembershipToGroupRetrieved $event)
     {
+        // Get from the cache the current user.
+        // If the users are different (use strcmp),
+        $key = sprintf('%s@getUsersThroughGroup:%s', UserGroup::class, $event->group->id());
+        if($this->cache->has($key)) {
+            $ids = collect($this->cache->get($key));
+            $newIds = collect($event->unionCloudUsers);
+            $updatingIds = $newIds->diff($ids);
+            $removingIds = $ids->diff($newIds);
+            $updatingIds->each(fn(int $id) => event(new UserAddedToGroup(
+                app(User::class)->getById($id),
+                $event->group
+            )));
+            $removingIds->each(fn(int $id) => event(new UserRemovedFromGroup(
+                app(User::class)->getById($id),
+                $event->group
+            )));
+        }
         $this->cache->forever(
-            sprintf('%s@getUsersThroughGroup:%s', UserGroup::class, $event->group->id()),
-            collect($event->unionCloudUsers)
+            $key,
+            collect($event->unionCloudUsers)->map(fn(DataUser $user) => $user->id())
         );
     }